# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SATOR OS Engine API
  version: "0.2.0"
  description: |
    Public HTTP/JSON API for the SATOR OS Engine. Optimization and reconstruction
    are asynchronous: POST to submit, then poll `/v1/jobs/{job_id}/result`.
    All requests except health and metrics must carry `x-api-key`.

    Acquisition function dispatch (multi-objective `min`/`max` only):
      - `qnehvi`      → qLogExpectedHypervolumeImprovement (default)
      - `qehvi`       → qExpectedHypervolumeImprovement
      - `qnoisyehvi`  → qNoisyExpectedHypervolumeImprovement
      - `parego`      → qLogExpectedImprovement on a Chebyshev-scalarized objective
      - `qei`         → qLogExpectedImprovement
      - `qpi`         → qProbabilityOfImprovement
      - `qucb`        → qUpperConfidenceBound

    Any problem whose objectives include advanced goal types (`target`,
    `within_range`, threshold goals, `explore`, `improve`) uses the Sobol +
    posterior-scoring path regardless of the `acquisition` string.
servers:
  - url: http://localhost:8080
    description: Local development (HTTP)
  - url: https://localhost:8443
    description: Local development (HTTPS)
security:
  - ApiKeyAuth: []
paths:
...
  /v1/jobs/{job_id}:
    get:
      summary: Get job status
      operationId: getJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current status of the job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
  /v1/jobs/{job_id}/result:
    get:
      summary: Get job result
      operationId: getJobResult
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Final result or failure details
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OptimizeJobResult'
                  - $ref: '#/components/schemas/ReconstructionJobResult'
                  - $ref: '#/components/schemas/JobFailure'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    JobStatus:
      type: object
      additionalProperties: true
      properties:
        job_id: { type: string }
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED]
        error:
          type: string
          description: Present when status=FAILED
      required: [job_id, status]
    JobFailure:
      type: object
      properties:
        status:
          type: string
          enum: [FAILED]
        error:
          type: string
      required: [status, error]
...
    Objectives:
      type: object
      description: Map of objective name to configuration
      additionalProperties:
        $ref: '#/components/schemas/ObjectiveConfig'
      minProperties: 1
    ObjectiveConfig:
      type: object
      additionalProperties: false
      properties:
        goal:
          type: string
          enum:
            - min
            - max
            - target
            - minimize_below
            - maximize_above
            - maximize_below
            - minimize_above
            - within_range
            - explore
            - improve
        target_value:
          type: number
          description: Required when goal=target
        threshold:
          $ref: '#/components/schemas/Threshold'
        range:
          $ref: '#/components/schemas/Range'
        weights:
          type: array
          items: { type: number }
          description: Optional; used for ParEGO scalarization
      required: [goal]
    Threshold:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          enum: ["<=", ">="]
        value: { type: number }
        weight:
          type: number
          description: Optional weighting on threshold satisfaction
      required: [type, value]
    Range:
      type: object
      additionalProperties: false
      properties:
        min: { type: number }
        max: { type: number }
        ideal:
          type: number
          description: Optional ideal target inside the range
        weight:
          type: number
          description: Range membership weight
        ideal_weight:
          type: number
          description: Weight on the ideal point closeness
      required: [min, max]
...
    ReconstructionJobResult:
      type: object
      additionalProperties: false
      properties:
        success: { type: boolean }
        reconstructed_formulation:
          type: object
          additionalProperties: false
          properties:
            ingredients: { type: array, items: { type: number } }
            parameters:  { type: array, items: { type: number } }
            combined:    { type: array, items: { type: number } }
            by_name:
              type: object
              additionalProperties: { type: number }
        reconstruction_metrics:
          type: object
          additionalProperties: true
          properties:
            final_error: { type: number }
            iterations:  { type: integer }
            method:      { type: string }
      required: [success]
    OptimizeJobResult:
      type: object
      additionalProperties: false
      properties:
        predictions:
          type: array
          items: { $ref: '#/components/schemas/Prediction' }
        pareto:
          type: object
          additionalProperties: false
          properties:
            indices:
              type: array
              items: { type: integer }
            points:
              type: array
              items:
                type: array
                items: { type: number }
          required: [indices, points]
        encoding_info:
          $ref: '#/components/schemas/EncodingInfo'
        diagnostics:
          $ref: '#/components/schemas/Diagnostics'
        objectives:
          $ref: '#/components/schemas/Objectives'
        encoded_dataset:
          type: array
          description: PCA-encoded dataset (if PCA used)
          items:
            type: array
            items: { type: number }
        gp_maps:
          $ref: '#/components/schemas/GPMaps'
      required: [predictions, pareto]
    Prediction:
      type: object
      additionalProperties: false
      properties:
        candidate:
          type: object
          description: Candidate values mapped back to original variable names
          additionalProperties:
            type: number
        objectives:
          type: array
          items: { type: number }
        variances:
          type: array
          items: { type: number }
        encoded:
          type: array
          description: PCA coordinates (if PCA used)
          items: { type: number }
        reconstructed:
          type: object
          description: Reconstruction (formulation) result when applicable
          additionalProperties: false
          properties:
            ingredients:
              type: array
              items: { type: number }
            parameters:
              type: array
              items: { type: number }
            combined:
              type: array
              items: { type: number }
            by_name:
              type: object
              additionalProperties: { type: number }
            success: { type: boolean }
            final_error: { type: number }
    EncodingInfo:
      type: object
      additionalProperties: false
      properties:
        pc_mins:
          type: array
          items: { type: number }
        pc_maxs:
          type: array
          items: { type: number }
        components:
          type: array
          items:
            type: array
            items: { type: number }
        mean:
          type: array
          items: { type: number }
    Diagnostics:
      type: object
      description: Implementation-defined diagnostics (device, GP details, fit metrics)
      additionalProperties: true
      properties:
        device: { type: string }
    GPMaps:
      type: object
      additionalProperties: false
      properties:
        space:
          type: string
          enum: [input, pca]
        dimension:
          type: integer
          enum: [2, 3]
        grid:
          type: object
          additionalProperties: false
          properties:
            axes:
              type: array
              items:
                type: array
                items: { type: number }
            resolution:
              type: array
              items: { type: integer, minimum: 2 }
              minItems: 2
              maxItems: 3
          required: [axes, resolution]
        maps:
          type: object
          additionalProperties: false
          properties:
            means:
              type: object
              additionalProperties:
                type: array
                description: 2D or 3D tensor flattened per axis order
                items:
                  oneOf:
                    - type: array
                      items:
                        type: array
                        items: { type: number }
                    - type: array
                      items:
                        type: array
                        items:
                          type: array
                          items: { type: number }
            variances:
              type: object
              additionalProperties:
                type: array
                items:
                  oneOf:
                    - type: array
                      items:
                        type: array
                        items: { type: number }
                    - type: array
                      items:
                        type: array
                        items:
                          type: array
                          items: { type: number }
          required: [means, variances]